Move base.cddl to blueprint and add golden LeiosVote test
