+++ .github/workflows/documentation.yml
@@ -27,6 +27,9 @@ jobs:
 
     steps:
       - uses: actions/checkout@v6
+        with:
+          submodules: true
+
       - uses: actions/setup-node@v6
         with:
           node-version: 24
