+++ ouroboros-consensus-cardano/test/cardano-test/Test/Consensus/Cardano/Golden.hs
@@ -2,11 +2,28 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Consensus.Cardano.Golden (tests) where
 
+import Cardano.Crypto.DSIGN (genKeyDSIGN, seedSizeDSIGN)
+import Cardano.Crypto.Seed (mkSeedFromBytes)
+import Cardano.Slotting.Slot (SlotNo (..))
+import qualified Data.ByteString as BS
+import Data.Data (Proxy (..))
+import Data.Word (Word8)
+import LeiosDemoTypes
+  ( EbHash (..)
+  , LeiosDSIGN
+  , LeiosPoint (..)
+  , LeiosSigningKey
+  , LeiosVote
+  , VoterId (MkVoterId)
+  , encodeLeiosVote
+  , signLeiosVote
+  )
 import Ouroboros.Consensus.Cardano.Block
 import Ouroboros.Consensus.Cardano.Node
 import Ouroboros.Consensus.HardFork.Combinator.Serialisation
@@ -22,17 +39,41 @@ import Test.Util.Serialisation.Golden
 
 tests :: TestTree
 tests =
-  goldenTest_all
-    codecConfig
-    ($(getGoldenDir) </> "cardano")
-    ( Just $
-        CDDLsForNodeToNode
-          ("ntnblock.cddl", "serialisedCardanoBlock")
-          ("ntnheader.cddl", "header")
-          ("ntntx.cddl", "tx")
-          ("ntntxid.cddl", "txId")
-    )
-    examples
+  testGroup
+    "Cardano"
+    [ goldenTest_all
+        codecConfig
+        ($(getGoldenDir) </> "cardano")
+        ( Just $
+            CDDLsForNodeToNode
+              ("ntnblock.cddl", "serialisedCardanoBlock")
+              ("ntnheader.cddl", "header")
+              ("ntntx.cddl", "tx")
+              ("ntntxid.cddl", "txId")
+        )
+        examples
+    , goldenTestCBOR
+        "LeiosVote"
+        typicalVote
+        encodeLeiosVote
+        ($(getGoldenDir) </> "cardano" </> "leios" </> "LeiosVote")
+        (Just ("leiosnotify.cddl", "vote"))
+    ]
+
+-- | A typical 'LeiosVote' at voter index 1000, signed with a deterministic
+-- key. Pinned by the corresponding golden file.
+typicalVote :: LeiosVote
+typicalVote =
+  signLeiosVote (mkLeiosSigningKey 0x42) (MkVoterId 1000) point
+ where
+  point = MkLeiosPoint (SlotNo 42) (MkEbHash (BS.pack [0 .. 31]))
+
+-- | Deterministic signing key, seeded by repeating a single byte.
+mkLeiosSigningKey :: Word8 -> LeiosSigningKey
+mkLeiosSigningKey b =
+  genKeyDSIGN $ mkSeedFromBytes $ BS.replicate sz b
+ where
+  sz = fromIntegral (seedSizeDSIGN (Proxy @LeiosDSIGN))
 
 instance
   CardanoHardForkConstraints c =>
